+++ gemrb/plugins/OpenALAudio/AmbientMgrAL.cpp
@@ -129,8 +129,8 @@ unsigned int AmbientMgrAL::tick(uint64_t ticks) const
 	int xpos, ypos;
 	core->GetAudioDrv()->GetListenerPos(xpos, ypos);
 	Point listener;
-	listener.x = (short) xpos;
-	listener.y = (short) ypos;
+	listener.x = xpos;
+	listener.y = ypos;
 
 	const Game* game = core->GetGame();
 	ieDword timeslice = 0;
